+++ src/assets/scripts/client/math/flightMath.js
@@ -86,37 +86,7 @@ export const getOffset = (aircraft, target, headingThruTarget = null) => {
     return offset;
 };
 
-// TODO: This has been replaced by `DynamicPositionModel.generateDynamicPositionFromBearingAndDistance()`.
-// Please replace all usages of this function with that, and then delete this helper function.
-/**
- * Get new position by fix-radial-distance method
- *
- * @param {array} fix       positional array of start point, in decimal-degrees [lat,lon]
- * @param {number} radial   heading to project along, in radians
- * @param {number} dist     distance to project, in nm
- * @returns {array}         location of the projected fix, in decimal-degrees [lat,lon]
- */
-export const fixRadialDist = (fix, radial, dist) => {
-    // TODO: if fix is a FixModel, there may already be a method for this. if there isnt there should be. `fix.positionInRadians`
-    // convert GPS coordinates to radians
-    fix = [
-        degreesToRadians(fix[0]),
-        degreesToRadians(fix[1])
-    ];
 
-    const R = PHYSICS_CONSTANTS.EARTH_RADIUS_NM;
-    // TODO: abstract these two calculations to functions
-    const lat2 = Math.asin(sin(fix[0]) * cos(dist / R) + cos(fix[0]) * sin(dist / R) * cos(radial));
-    const lon2 = fix[1] + Math.atan2(
-        sin(radial) * sin(dist / R) * cos(fix[0]),
-        cos(dist / R) - sin(fix[0]) * sin(lat2)
-    );
-
-    return [
-        radiansToDegrees(lat2),
-        radiansToDegrees(lon2)
-    ];
-};
 
 /**
  *